+++ src/features/kyc/home/components/sections/HeroSection.tsx
