+++ src/features/kyc/home/components/sections/InvestmentSection.tsx
