+++ src/features/kyc/home/components/sections/InvestorSection.tsx
