+++ src/features/kyc/home/components/sections/MediaSection.tsx
