+++ src/features/kyc/home/hooks/useHomeStaticQuery.ts
